+++ src/core/objective.jl
@@ -16,38 +16,10 @@ minimum load delta objective (continuous load shed) with storage
 ```
 """
 function objective_mc_min_load_setpoint_delta_switch(pm::PMD.AbstractUnbalancedPowerModel)
-    for (n, nw_ref) in PMD.nws(pm)
-        PMD.var(pm, n)[:delta_pg] = Dict(
-            i => PMD.JuMP.@variable(
-                pm.model,
-                [c in PMD.ref(pm, n, :gen, i)["connections"]],
-                base_name="$(n)_$(i)_delta_pg",
-                start = 0.0
-            ) for i in PMD.ids(pm, n, :gen)
-        )
+    nw_ids = sort(collect(PMD.nw_ids(pm)))
 
-        for (i, gen) in nw_ref[:gen]
-            for (idx, c) in enumerate(gen["connections"])
-                PMD.JuMP.@constraint(pm.model, PMD.var(pm, n, :delta_pg, i)[c] >=  (gen["pg"][idx] - PMD.var(pm, n, :pg, i)[c]))
-                PMD.JuMP.@constraint(pm.model, PMD.var(pm, n, :delta_pg, i)[c] >= -(gen["pg"][idx] - PMD.var(pm, n, :pg, i)[c]))
-            end
-        end
-
-        PMD.var(pm, n)[:delta_ps] = Dict(
-            i => PMD.JuMP.@variable(
-                pm.model,
-                [c in PMD.ref(pm, n, :storage, i)["connections"]],
-                base_name="$(n)_$(i)_delta_ps",
-                start = 0.0
-            ) for i in PMD.ids(pm, n, :storage)
-        )
-
-        for (i, strg) in nw_ref[:storage]
-            for (idx, c) in enumerate(strg["connections"])
-                PMD.JuMP.@constraint(pm.model, PMD.var(pm, n, :delta_ps, i)[c] >=  (strg["ps"][idx] - PMD.var(pm, n, :ps, i)[c]))
-                PMD.JuMP.@constraint(pm.model, PMD.var(pm, n, :delta_ps, i)[c] >= -(strg["ps"][idx] - PMD.var(pm, n, :ps, i)[c]))
-            end
-        end
+    for (i, n) in enumerate(nw_ids)
+        nw_ref = PMD.ref(pm, n)
 
         PMD.var(pm, n)[:delta_sw_state] = PMD.JuMP.@variable(
             pm.model,
@@ -56,22 +28,29 @@ function objective_mc_min_load_setpoint_delta_switch(pm::PMD.AbstractUnbalancedP
             start = 0
         )
 
-        for (i,switch) in nw_ref[:switch_dispatchable]
-            PMD.JuMP.@constraint(pm.model, PMD.var(pm, n, :delta_sw_state, i) >=  (switch["state"] - PMD.var(pm, n, :switch_state, i)))
-            PMD.JuMP.@constraint(pm.model, PMD.var(pm, n, :delta_sw_state, i) >= -(switch["state"] - PMD.var(pm, n, :switch_state, i)))
+        for (s,switch) in nw_ref[:switch_dispatchable]
+            z_switch = PMD.var(pm, n, :switch_state, s)
+            if i == 1
+                PMD.JuMP.@constraint(pm.model, PMD.var(pm, n, :delta_sw_state, s) >=  (PMD.JuMP.start_value(z_switch) - z_switch))
+                PMD.JuMP.@constraint(pm.model, PMD.var(pm, n, :delta_sw_state, s) >= -(PMD.JuMP.start_value(z_switch) - z_switch))
+            else  # multinetwork
+                z_switch_prev = PMD.var(pm, nw_ids[i-1], :switch_state, s)
+                PMD.JuMP.@constraint(pm.model, PMD.var(pm, n, :delta_sw_state, s) >=  (z_switch_prev - z_switch))
+                PMD.JuMP.@constraint(pm.model, PMD.var(pm, n, :delta_sw_state, s) >= -(z_switch_prev - z_switch))
+            end
         end
     end
 
-    w = Dict(n => Dict(i => 100*get(load, "weight", 1.0) for (i,load) in PMD.ref(pm, n, :load)) for n in PMD.nw_ids(pm))
+    # weight for discouraging switch state changes should be an order of magnitude
+    # smaller than the smallest block weight (to ensure all blocks get restored)
+    Wdsw = Dict(n => minimum(filter(x->x==0,collect(values(PMD.ref(pm, n, :block_weights))))) for n in PMD.nw_ids(pm))
 
     PMD.JuMP.@objective(pm.model, Min,
         sum(
-            sum(      10*(1 - PMD.var(pm, n, :z_voltage, i)) for  (i,bus) in   nw_ref[:bus]) +
-            sum( w[n][i]*(1 - PMD.var(pm, n, :z_demand, i)) for  (i,load) in  nw_ref[:load]) +
-            sum(         (1 - PMD.var(pm, n, :z_shunt, i)) for  (i,shunt) in nw_ref[:shunt]) +
-            sum( 1e-4 * sum(PMD.var(pm, n, :delta_pg, i)[c] for (idx,c) in enumerate(gen["connections"])) for (i,gen)  in nw_ref[:gen]) +
-            sum( 1e-4 * sum(PMD.var(pm, n, :delta_ps, i)[c] for (idx,c) in enumerate(strg["connections"])) for (i,strg) in nw_ref[:storage]) +
-            sum( 1e-3 * sum(PMD.var(pm, n, :delta_sw_state, l)) for l in PMD.ids(pm, n, :switch_dispatchable))
+            sum( 1e-2 * sum(gen["cost"][1] * PMD.var(pm, n, :pg, i)[c] + gen["cost"][2] for c in gen["connections"]) for (i,gen) in nw_ref[:gen]) +
+            sum( 1e3  * PMD.ref(pm, n, :block_weights, i)*(1-PMD.var(pm, n, :z_block, i)) for (i,block) in nw_ref[:blocks]) +
+            sum( 1e-1  * PMD.ref(pm, n, :switch_scores, l)*(1-PMD.var(pm, n, :switch_state, l)) for l in PMD.ids(pm, n, :switch_dispatchable) ) +
+            sum( 1e-4 * Wdsw[n] * sum(PMD.var(pm, n, :delta_sw_state, l)) for l in PMD.ids(pm, n, :switch_dispatchable))
         for (n, nw_ref) in PMD.nws(pm))
     )
 end